backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
email-validator==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1